mysql/server: Provide NextConnectionID hook on Listener
This is bringing forward: https://github.com/mattrobenolt/vitess/commit/2ee1752603d89391029a1bdff614debfe59e2c97

We use this in edge-gateway to change the connection ID generation to
not be a global sequence.

I also slightly changed the implementation to not take a pointer, but to
just pass the uint32 and return a new uint32 by value instead. We don't
need this to be a pointer anymore.